refactor: separate appTheme from main
